+++ server/extension/sql/60_postalcodes.sql
@@ -34,6 +34,12 @@ RETURNS Geometry AS $$
       quota_service.increment_total_service_use()
 $$ LANGUAGE plpythonu;
 
+CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_point(username text, orgname text, code double precision)
+RETURNS Geometry AS $$
+  SELECT cdb_dataservices_server.cdb_geocode_postalcode_point(username, orgname, code::text);
+$$ LANGUAGE SQL;
+
+
 CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_point(username text, orgname text, code text, country text)
 RETURNS Geometry AS $$
   from cartodb_services.metrics import metrics
@@ -69,6 +75,12 @@ RETURNS Geometry AS $$
     finally:
       quota_service.increment_total_service_use()
 $$ LANGUAGE plpythonu;
+
+CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_point(username text, orgname text, code double precision, country text)
+RETURNS Geometry AS $$
+  SELECT cdb_dataservices_server.cdb_geocode_postalcode_point(username, orgname, code::text, country);
+$$ LANGUAGE SQL;
+
 
 CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_polygon(username text, orgname text, code text)
 RETURNS Geometry AS $$
@@ -106,6 +118,13 @@ RETURNS Geometry AS $$
       quota_service.increment_total_service_use()
 $$ LANGUAGE plpythonu;
 
+CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_polygon(username text, orgname text, code double precision)
+RETURNS Geometry AS $$
+  SELECT cdb_dataservices_server.cdb_geocode_postalcode_polygon(username, orgname, code::text)
+$$ LANGUAGE SQL;
+
+
+
 CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_polygon(username text, orgname text, code text, country text)
 RETURNS Geometry AS $$
   from cartodb_services.metrics import metrics
@@ -141,6 +160,11 @@ RETURNS Geometry AS $$
     finally:
       quota_service.increment_total_service_use()
 $$ LANGUAGE plpythonu;
+
+CREATE OR REPLACE FUNCTION cdb_dataservices_server.cdb_geocode_postalcode_polygon(username text, orgname text, code double precision, country text)
+RETURNS Geometry AS $$
+  SELECT cdb_dataservices_server.cdb_geocode_postalcode_polygon(username, orgname, code::text, country);
+$$ LANGUAGE SQL;
 
 --------------------------------------------------------------------------------